+++ website/vpn/sts/forms.py
@@ -71,34 +71,56 @@ class AddForm(Form):
     start_type = SelectField(u'启动类型',
                              choices=[('add', u'手工连接'), ('start', u'服务启动自动连接')])
 
-    encryption_algorithm = SelectField(u'加密算法',
-                                       choices=[('3des', u'3des'), ('cast128', u'cast128'),
-                                                ('blowfish128', u'blowfish128'), ('blowfish192', u'blowfish192'),
-                                                ('blowfish256', u'blowfish256'), ('aes128', u'aes128'),
-                                                ('aes192', u'aes192'), ('aes256', u'aes256'),
-                                                ('camellia128', u'camellia128'), ('camellia192', u'camellia192'),
-                                                ('camellia256', u'camellia256'), ('null', u'不加密')])
+    ike_encryption_algorithm = SelectField(u'IKEv2 加密算法',
+                                           choices=[('3des', u'3DES'), ('aes128', u'AES128'),
+                                                    ('aes192', u'AES192'), ('aes256', u'AES256')])
 
-    integrity_algorithm = SelectField(u'签名算法',
-                                      choices=[('md5', u'md5'), ('md5_128', u'md5_128'),
-                                               ('sha1', u'sha1'), ('sha1_160', u'sha1_160'),
-                                               ('aesxcbc', u'aesxcbc'), ('aescmac', u'aescmac'),
-                                               ('aes128gmac', u'aes128gmac'), ('aes192gmac', u'aes192gmac'),
-                                               ('aes256gmac', u'aes256gmac'), ('sha2_256', u'sha2_256'),
-                                               ('sha2_384', u'sha2_384'), ('sha2_512', u'sha2_512'),
-                                               ('sha2_256_96', u'sha2_256_96')])
+    ike_integrity_algorithm = SelectField(u'IKEv2 验证算法',
+                                          choices=[('md5', u'MD5'), ('md5_128', u'MD5-128'),
+                                                   ('sha1', u'SHA1'), ('sha1_160', u'SHA1-160'),
+                                                   ('sha2_256', u'SHA2-256'), ('sha2_384', u'SHA2-384'),
+                                                   ('sha2_512', u'SHA2-512'), ('sha2_256_96', u'SHA2-256-96'),
+                                                   ('aesxcbc', u'AES-XCBC'), ('aescmac', u'AES-CMAC'),
+                                                   ('aes128gmac', u'AES-128-GMAC'), ('aes192gmac', u'AES-192-GMAC'),
+                                                   ('aes256gmac', u'AES-256-GMAC')])
 
-    dh_algorithm = SelectField(u'DH算法',
-                               choices=[('modp768', u'modp768(group 1)'), ('modp1024', u'modp1024(group 2)'),
-                                        ('modp1536', u'modp1536(group 5)'), ('modp2048', u'modp2048(group 14)'),
-                                        ('modp3072', u'modp3072(group 15)'), ('modp4096', u'modp4096(group 16)'),
-                                        ('modp6144', u'modp6144(group 17)'), ('modp8192', u'modp8192(group 18)'),
-                                        ('modp1024s160', u'modp1024s160(group 22)'), ('modp2048s224', u'modp2048s224(group 23)'),
-                                        ('modp2048s256', u'modp2048s256(group 24)'), ('ecp192', u'ecp192(group 25)'),
-                                        ('ecp224', u'ecp224(group 26)'), ('ecp256', u'ecp256(group 19)'),
-                                        ('ecp384', u'ecp384(group 20)'), ('ecp521', u'ecp521(group 21)'),
-                                        ('ecp224bp', u'ecp224bp(group 27)'), ('ecp256bp', u'ecp256bp(group 28)'),
-                                        ('ecp384bp', u'ecp384bp(group 29)'), ('ecp512bp', u'ecp512bp(group 30)')])
+    ike_dh_algorithm = SelectField(u'IKEv2 DH 组',
+                                   choices=[('modp768', u'Group 1 modp768'), ('modp1024', u'Group 2 modp1024'),
+                                            ('modp1536', u'Group 5 modp1536'), ('modp2048', u'Group 14 modp2048'),
+                                            ('modp3072', u'Group 15 modp3072'), ('modp4096', u'Group 16 modp4096'),
+                                            ('modp6144', u'Group 17 modp6144'), ('modp8192', u'Group 18 modp8192'),
+                                            ('ecp256', u'Group 19 ecp256'), ('ecp384', u'Group 20 ecp384'),
+                                            ('ecp521', u'Group 21 ecp521'), ('modp1024s160', u'Group 22 modp1024s160'),
+                                            ('modp2048s224', u'Group 23 modp2048s224'), ('modp2048s256', u'Group 24 modp2048s256'),
+                                            ('ecp192', u'Group 25 ecp192'), ('ecp224', u'Group 26 ecp224'),
+                                            ('ecp224bp', u'Group 27 ecp224bp'), ('ecp256bp', u'Group 28 ecp256bp'),
+                                            ('ecp384bp', u'Group 29 ecp384bp'), ('ecp512bp', u'Group 30 ecp512bp')])
+
+    esp_encryption_algorithm = SelectField(u'ESP 加密算法',
+                                           choices=[('3des', u'3DES'), ('aes128', u'AES128'),
+                                                    ('aes192', u'AES192'), ('aes256', u'AES256')])
+
+    esp_integrity_algorithm = SelectField(u'ESP 验证算法',
+                                          choices=[('md5', u'MD5'), ('md5_128', u'MD5-128'),
+                                                   ('sha1', u'SHA1'), ('sha1_160', u'SHA1-160'),
+                                                   ('sha2_256', u'SHA2-256'), ('sha2_384', u'SHA2-384'),
+                                                   ('sha2_512', u'SHA2-512'), ('sha2_256_96', u'SHA2-256-96'),
+                                                   ('aesxcbc', u'AES-XCBC'), ('aescmac', u'AES-CMAC'),
+                                                   ('aes128gmac', u'AES-128-GMAC'), ('aes192gmac', u'AES-192-GMAC'),
+                                                   ('aes256gmac', u'AES-256-GMAC')])
+
+    esp_dh_algorithm = SelectField(u'ESP DH 组',
+                                   choices=[('null', u'无'),
+                                            ('modp768', u'Group 1 modp768'), ('modp1024', u'Group 2 modp1024'),
+                                            ('modp1536', u'Group 5 modp1536'), ('modp2048', u'Group 14 modp2048'),
+                                            ('modp3072', u'Group 15 modp3072'), ('modp4096', u'Group 16 modp4096'),
+                                            ('modp6144', u'Group 17 modp6144'), ('modp8192', u'Group 18 modp8192'),
+                                            ('ecp256', u'Group 19 ecp256'), ('ecp384', u'Group 20 ecp384'),
+                                            ('ecp521', u'Group 21 ecp521'), ('modp1024s160', u'Group 22 modp1024s160'),
+                                            ('modp2048s224', u'Group 23 modp2048s224'), ('modp2048s256', u'Group 24 modp2048s256'),
+                                            ('ecp192', u'Group 25 ecp192'), ('ecp224', u'Group 26 ecp224'),
+                                            ('ecp224bp', u'Group 27 ecp224bp'), ('ecp256bp', u'Group 28 ecp256bp'),
+                                            ('ecp384bp', u'Group 29 ecp384bp'), ('ecp512bp', u'Group 30 ecp512bp')])
 
     local_subnet = TextAreaField(u'本端子网',
                                  validators=[DataRequired(message=u'这是一个必选项！'),