+++ website/vpn/sts/services.py
@@ -253,12 +253,16 @@ def vpn_settings(form, tunnel_id=None):
     vpn = VpnServer()
     local_subnet = ','.join([i.strip() for i in form.local_subnet.data.split(',')])
     remote_subnet = ','.join([i.strip() for i in form.remote_subnet.data.split(',')])
-    esp = '%s-%s-%s' % (form.encryption_algorithm.data, form.integrity_algorithm.data, form.dh_algorithm.data)
-    rules = {'auto': form.start_type.data, 'esp': esp,
-             'left': '0.0.0.0', 'leftsubnet': local_subnet,
+    if form.esp_dh_algorithm.data == 'null':
+        esp = '%s-%s' % (form.esp_encryption_algorithm.data, form.esp_integrity_algorithm.data)
+    else:
+        esp = '%s-%s-%s' % (form.esp_encryption_algorithm.data, form.esp_integrity_algorithm.data, form.esp_dh_algorithm.data)
+    ike = '%s-%s-%s' % (form.ike_encryption_algorithm.data, form.ike_integrity_algorithm.data, form.ike_dh_algorithm.data)
+    rules = {'left': '0.0.0.0', 'leftsubnet': local_subnet,
              'leftid': form.tunnel_name.data, 'right': form.remote_ip.data,
              'rightsubnet': remote_subnet, 'rightid': form.tunnel_name.data,
-             'authby': 'secret'}
+             'authby': 'secret', 'esp': esp,
+             'ike': ike, 'auto': form.start_type.data}
     if tunnel.update_tunnel(tunnel_id, form.tunnel_name.data, json.dumps(rules),
                             form.psk.data) and vpn.reload:
         return True